[FIXED JENKINS-5492] Hide & encrypt Bind Password.
Hide the Bind Password by changing text field to password field in config.jelly.
Encrypt the Bind Password by changing password type from String to Secret.
Upgrade the core version to 1.436 in pom.xml file to support JDK 7 or higher and
to use jenkinsRule in the LdapSearchTest.
Add jenkinsRule and annotations for the tests to solve the NullPointer Exception
caused by password type changing.
When the user upgrades the plugin, they need to click save in the config page in
order to encrypt the password on the file system.